Introduce update_author_term(), which conveniently deprecates refresh_coauthor_term()
This update_author_term() mechanism handles creating the author term if it doesn't yet exist, prefixing the existing term if it's not yet prefixed, and updating the term description with the latest coauthor object details.

In fact, consolidating our term updating to this single method should also fix our issue with colliding with other taxonomies, as we only care about collisions if we're creating a new term.

See #65