Allow guest authors to be created from users, even when they may not have display_name, first_name or last_name fields.
We're doing our best to avoid the situation where a guest author is created with the same user_login as a user. In this scenario, they share the same author term which means that any modifications to one lead to modifications to the other. Where this is the biggest possibility is if the guest author becomes unmapped to the user, posts are continued to be assigned to it, and then the guest author is deleted.

So, the easiest thing to do is restrict the mapping from being changed

See #28, #30