Rewrite the 'user_has_cap' filter for clarity.
We want to apply the 'edit_others_posts' cap for a user when they're a co-author of a post. Depending on the status of
 the post, we may also want to apply additional caps if they have them.

We apply the 'edit_others_posts' cap when filtering the 'edit_post' cap for a post object. Due to a couple of deficiencies in core, we also need to apply it a couple more times

Fixes #61, #8