Providing arguments for optional span class
As automatically adding the span class might break existing themes,
moved this implementation to make it optional and applied through
filter, while retaining the in-link class.

I'm unsure of coding standards for this setup, but I figured that the
bracketing variables should be added after the sprintf and it made sense
to break down the return into two steps. Open to a better way of doing
this if there is one.