Model core's behavior when determining whether the author archive should be displayed
The author archive should be displayed when the user is a member of the blog, or the user has a previously published post on the blog.

The author archive shouldn't ever be displayed when the user has no published posts on the blog.

See https://core.trac.wordpress.org/changeset/27290

Fixes #133, #169