Display Proper Author Name in Post Archive Links
Previous functionality relied on `get_the_author()`, which relies on `$authordata` which may hold the data for the primary author rather than the appropriate coauthor. The coauthor link was still correct, however.

Utilizing `$author->display_name` instead ensures that the proper author name is displayed along with the coauthor link.