Add better Infinite Scroll support for Guest Authors
Filter the IS settings to remove any author ids and replace them with
`author_name`, so that the standard CAP Guest Authors query filters
work as expected.

Problem existed when a Guest Author slug matched a real user in the
network (not necessarily a member of the site), and WP injects `author`
into the query vars.

IS would pick that up for future requests, which ended up with queries
that CAP didn’t properly understand / fix.

Fixes #229