Improve CLI commands with more flexible author lookups
Attempts both a lookup on both the unsanitized and sanitized version of
`user_login` when assigning co-authors, for imported content that may
have used a different user system in the past, with different u/n rules.

Also rectifies the username passed to actual add_coauthors() call when
making the assignment. Previously we passed the `$original_author` -
now we pass in the `user_nicename` of the object retrieved from the
database, as it doesn’t necessarily match `$original_author`