Route arguments big refactor.
Since there was a lot of duplication inside the `get_args()` method, the structure was changed from the inside out.
Instead of defining parent contexts, it now defines the fields. A new method `filter_args` was added to help fetch
this arguments.

This cleans up the route args and has no duplicaiton whatsoever.