Add filter to get_coauthors
Adds a filter to get_coauthors to allow hooking in and providing coauthor objects that might not exist in the db. Use case: we have several places in our site where we output authors based on get_coauthors (templates, some custom POST requests we send out to some API's, etc). We have the need to "fake" some coauthors in some instances, and want to be able to dynamically insert them in when a post meets certain criteria. Being able to filter the data that gets returned when calling get_coauthors will allow us to take advantage of all existing code where we use get_coauthors, but will allow us to dynamically inject "fake" authors when needed.