Fix test_coauthors_links_single_when_guets_author(), which was wrongly passing.
The test passed because was updating a postmeta instead of the user record in wp_users, thus subsequent assertions all passed simply because no real edit had been done.
When making the edit for real, the test would fail in several places, so tweaks were needed to fix its behavior.