Modified query filter regex.
This code is meant to append a conditional statement to the WHERE clause of a query.

But because the regex mistakenly believed that the statement should be appended after a user ID, and not after 'private', we ended up with queries like this:

```
OR wp_posts.post_author = 3
OR (
    wp_term_taxonomy.taxonomy = 'author'
    AND wp_term_taxonomy.term_id = '8'
)
AND wp_posts.post_status = 'private'
```

instead of this:

```
OR wp_posts.post_author = 3 AND wp_posts.post_status = 'private'
OR (
    wp_term_taxonomy.taxonomy = 'author'
    AND wp_term_taxonomy.term_id = '8'
)
```

Those are two very different queries: One returns posts including those private posts that belong to user 3, the other excludes all posts that are not private posts belonging to user 3.

The modified regex now allows the query to be constructed appropriately.

I've run unit tests on the code and no regressions were reported.

You can also review my process of solving the problem here: https://github.com/Automattic/Co-Authors-Plus/issues/508