Modify test so that query is run as post author, not as admin.
An issue was raised (https://github.com/Automattic/Co-Authors-Plus/issues/508) that was not caught by existing unit tests because it only presented itself when the query was performed as an author. The existing unit test performed the query as an admin, which caused the issue to not appear, and the test to pass.

Changing the test to run the query in the context of the post author should prevent this from happening again.