Change permission
For many newsrooms, including ours, we have users that we don't want to have edit access. For example, our photographers have accounts because they use our backend to submit their photos. Sometimes though, we want to assign them as the author of a post. For example, if they make a photo essay.