Fix `undefined index: post` errors
Return `$post_data` in `theaction_bulk_edit_update_coauthors` function to prevent undefined errors.

`wp_insert_post_data` is a filter so we need to return a value. When we aren't bulk editing, the `action_bulk_edit_update_coauthors` function still fires on every post data insert, so we need to return the value of the post data when we exit it early in L621 to prevent the undefined warnings.

Also simplified the text instructions for the author bulk edit to emphasize that all previous authors are removed when new authors are added.