fix: prevent the backfill from running forever. (#1065)
* fix: prevent the backfill from running forever.

There's an edge case where an author that no longer exists can still be assigned to a post. This throws the backfill script into an infinite loop, because the respective author-term is never found/created, and so the underlying problem of missing author-term records is never resolved. The infinite loop is started when at the end of the while loop, the script asks for "remaining posts which need author terms" and so it returns the same rows over and over.

This fix addresses this in 2 ways:
1. If an author is not found, we look for the most prolific author on the site and assign the posts to them. If there is no prolific author, one is created. And if one can't be created, an exception is thrown so that the script can't proceed.
2. Checks have been added so that the script can't go beyond what should be the maximum number of rows needing to be addressed.

* fix: obtaining the first available admin user account instead.

* fix: updating output to reflect that the ID belongs to an Admin account.

* fix: this function should be private

* fix: switching tactic to skipping posts that have missing post_author.

This approach is more faithful with what the current condition on the site would be anyway. If the post author doesn't exist on the site, you wouldn't be able to see the particular post in question in an author archive anyway. Skipping the post instead of reassigning it to the first available admin user is a cleaner solution.

* fix: removed unused references from a past commit

* fix: appeasing PHPCS